moved deployer to own file. cleanup
